+++ spring-cloud-alibaba-starters/spring-cloud-starter-dubbo/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 com.alibaba.cloud.dubbo.autoconfigure.DubboMetadataAutoConfiguration,\
+com.alibaba.cloud.dubbo.autoconfigure.DubboMetadataBeanConfiguration,\
 com.alibaba.cloud.dubbo.autoconfigure.DubboOpenFeignAutoConfiguration,\
 com.alibaba.cloud.dubbo.autoconfigure.DubboServiceRegistrationAutoConfiguration,\
 com.alibaba.cloud.dubbo.autoconfigure.DubboServiceRegistrationNonWebApplicationAutoConfiguration,\
